rename directory to match package name
