+++ components/movementsensor/register/register.go
@@ -7,7 +7,7 @@ import (
 	_ "go.viam.com/rdk/components/movementsensor/dualgps"
 	_ "go.viam.com/rdk/components/movementsensor/fake"
 	_ "go.viam.com/rdk/components/movementsensor/gpsnmea"
-	_ "go.viam.com/rdk/components/movementsensor/gpsrtkserial"
+	_ "go.viam.com/rdk/components/movementsensor/gpsrtk"
 	_ "go.viam.com/rdk/components/movementsensor/imuvectornav"
 	_ "go.viam.com/rdk/components/movementsensor/imuwit"
 	_ "go.viam.com/rdk/components/movementsensor/merged"
